Partially Implemented `BackupService`

+++ Libraries/Core/Utils/FileUtil.cs
@@ -1,9 +1,11 @@
+using Core.Models.Utils;
 using SptCommon.Annotations;
 
 namespace Core.Utils;
 
 [Injectable]
-public class FileUtil
+public class FileUtil(
+    ISptLogger<FileUtil> _logger)
 {
     public List<string> GetFiles(string path, bool recursive = false)
     {
@@ -50,7 +52,6 @@ public class FileUtil
         Directory.CreateDirectory(path);
     }
 
-
     public bool FileExists(string path)
     {
         return File.Exists(path);
@@ -77,6 +78,35 @@ public class FileUtil
 
     public void DeleteFile(string filePath)
     {
+        if (!FileExists(filePath))
+        {
+            _logger.Error($"Unable to delete file, not found: {filePath}");
+
+            return;
+        }
+
         File.Delete(filePath);
+    }
+
+    /// <summary>
+    /// Copy a file from one path to another
+    /// </summary>
+    /// <param name="copyFromPath">Source file to copy from</param>
+    /// <param name="destinationPath"></param>
+    /// <param name="overwrite">Should destination file be overwritten</param>
+    /// <exception cref="NotImplementedException"></exception>
+    public void CopyFile(string copyFromPath, string destinationPath, bool overwrite = false)
+    {
+        // Check it exists first
+        if (!FileExists(copyFromPath))
+        {
+            _logger.Error($"Source file not found: {copyFromPath}. Cannot copy to: {destinationPath}");
+        }
+
+        // Ensure dir exists
+        Directory.CreateDirectory(destinationPath);
+
+        // Copy the file
+        File.Copy(copyFromPath, destinationPath, overwrite);
     }
 }
